fix(HQM-248): src/models/index.ts -> src/models/contact.ts -> src/helpers/await-for-realm.ts
